+++ src/main/java/com/api/ttoklip/domain/question/main/dto/response/QuestionCategoryResponse.java
@@ -1,4 +1,4 @@
-package com.api.ttoklip.domain.question.dto.response;
+package com.api.ttoklip.domain.question.main.dto.response;
 
 public record QuestionCategoryResponse(QuestionResponse mainCategoryData) {
 }
