+++ src/main/java/com/api/ttoklip/domain/question/main/dto/response/QuestionSearchResponse.java
@@ -1,4 +1,4 @@
-package com.api.ttoklip.domain.question.dto.response;
+package com.api.ttoklip.domain.question.main.dto.response;
 
 public record QuestionSearchResponse(QuestionResponse searchData,  Integer totalPage,
                                      Long totalElements, Boolean isFirst, Boolean isLast) {
